NewAsm: fixed indexes in loops

+++ src/new-asm/asm-program.h
@@ -26,7 +26,7 @@ public:
 
   int getFunciontsTotalSize() {
     int totSize = 0;
-    for(int i = 0; i < functions.size(); i++) {
+    for(size_t i = 0; i < functions.size(); i++) {
       totSize += functions[i]->getSize();
     }
     return totSize;
